requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
attrs==23.2.0
    # via wmctrl
certifi==2024.2.2
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
distlib==0.3.8
    # via virtualenv
fancycompleter==0.9.1
    # via pdbpp
filelock==3.14.0
    # via virtualenv
identify==2.5.36
    # via pre-commit
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
nodeenv==1.9.0
    # via pre-commit
packaging==24.0
    # via pytest
pdbpp==0.10.3
    # via -r requirements.in
platformdirs==4.2.2
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.7.1
    # via -r requirements.in
pydantic==2.7.2
    # via -r requirements.in
pydantic-core==2.18.3
    # via pydantic
pygments==2.18.0
    # via pdbpp
pyrepl==0.9.0
    # via fancycompleter
pytest==8.2.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
requests==2.32.3
    # via -r requirements.in
text-unidecode==1.3
    # via python-slugify
tqdm==4.66.4
    # via -r requirements.in
typing-extensions==4.12.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
virtualenv==20.26.2
    # via pre-commit
wmctrl==0.5
    # via pdbpp
